Refactoring NewGameIntentHandler to attempt to use Quiz object in SessionAttributes

+++ src/main/groovy/com/lorenjamison/alexa/triviaroyale/dataobject/base/CategoryBase.groovy
@@ -1,6 +1,6 @@
 package com.lorenjamison.alexa.triviaroyale.dataobject.base
 
 class CategoryBase {
-    int id
+    long id
     String name
 }
